Combine StaccatoAccessBarrier into RealtimeAccessBarrier
