+++ runtime/CMakeLists.txt
@@ -335,9 +335,6 @@ endif()
 if(J9VM_MODULE_GC_REALTIME)
 	add_subdirectory(gc_realtime)
 endif()
-if(J9VM_MODULE_GC_STACCATO)
-	add_subdirectory(gc_staccato)
-endif()
 if(J9VM_MODULE_GC_STATS)
 	add_subdirectory(gc_stats)
 endif()
