+++ runtime/gc_realtime/RealtimeAccessBarrier.hpp
@@ -1,6 +1,5 @@
-
 /*******************************************************************************
- * Copyright (c) 1991, 2014 IBM Corp. and others
+ * Copyright (c) 1991, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -32,8 +31,10 @@
 
 #include "Metronome.hpp"
 #include "ObjectAccessBarrier.hpp"
+#include "RememberedSetSATB.hpp"
 
 class MM_EnvironmentBase;
+class MM_EnvironmentRealtime;
 class MM_RealtimeMarkingScheme;
 class MM_RealtimeGC;
 
@@ -54,8 +55,10 @@ public:
 protected:
 	MM_RealtimeMarkingScheme *_markingScheme;	
 	MM_RealtimeGC *_realtimeGC;
+
 private:
-	
+	bool _doubleBarrierActive; /**< Global indicator that the double barrier is active. New threads will be set to double barrier mode if this falg is true. */
+
 /* Methods */
 public:
 	/* Constructors & destructors */
@@ -66,13 +69,13 @@ public:
 	{
 		_typeId = __FUNCTION__;
 	}
-	
+
 	/* Inherited from MM_ObjectAccessBarrier */
 	virtual J9Object* referenceGet(J9VMThread *vmThread, J9Object *refObject);
 	virtual void jniDeleteGlobalReference(J9VMThread *vmThread, J9Object *reference);
 	virtual void stringConstantEscaped(J9VMThread *vmThread, J9Object *stringConst);
 	virtual void deleteHeapReference(MM_EnvironmentBase *env, J9Object *object);
-	
+
 	virtual void storeObjectToInternalVMSlot(J9VMThread *vmThread, J9Object** destSlot, J9Object *value);
 
 	virtual void* jniGetPrimitiveArrayCritical(J9VMThread* vmThread, jarray array, jboolean *isCopy);
@@ -97,18 +100,66 @@ protected:
 	/* Constructors & destructors */
 	virtual bool initialize(MM_EnvironmentBase *env);
 	virtual void tearDown(MM_EnvironmentBase *env);
-	
+	virtual void kill(MM_EnvironmentBase *env);
+
 	/* Inherited from MM_ObjectAccessBarrier */
 	virtual mm_j9object_t readObjectFromInternalVMSlotImpl(J9VMThread *vmThread, j9object_t *srcAddress, bool isVolatile=false);
-	
+
 	/* New methods */
 	void validateWriteBarrier(J9VMThread *vmThread, J9Object *dstObject, fj9object_t *dstAddress, J9Object *srcObject);
-	virtual void rememberObjectImpl(MM_EnvironmentBase *env, J9Object *object) = 0;
-	
+	virtual void rememberObjectImpl(MM_EnvironmentBase *env, J9Object *object);
+
 	/* New methods */
 	void rememberObject(MM_EnvironmentBase *env, J9Object *object);
 	void rememberObjectIfBarrierEnabled(J9VMThread *vmThread, J9Object* object);
-	
+
+public:
+	static MM_RealtimeAccessBarrier *newInstance(MM_EnvironmentBase *env);
+	MMINLINE void setDoubleBarrierActive() { _doubleBarrierActive = true; }
+	MMINLINE void setDoubleBarrierInactive() { _doubleBarrierActive = false; }
+	MMINLINE bool isDoubleBarrierActive() { return _doubleBarrierActive; }
+	void setDoubleBarrierActiveOnThread(MM_EnvironmentBase* env);
+	void setDoubleBarrierInactiveOnThread(MM_EnvironmentBase* env);
+	virtual void initializeForNewThread(MM_EnvironmentBase* env);
+
+	virtual bool preObjectStore(J9VMThread *vmThread, J9Object *destObject, fj9object_t *destAddress, J9Object *value, bool isVolatile=false);
+	virtual bool preObjectStore(J9VMThread *vmThread, J9Object *destClass, J9Object **destAddress, J9Object *value, bool isVolatile=false);
+	virtual bool preObjectStore(J9VMThread *vmThread, J9Object **destAddress, J9Object *value, bool isVolatile=false);
+
+	virtual I_32 backwardReferenceArrayCopyIndex(J9VMThread *vmThread, J9IndexableObject *srcObject, J9IndexableObject *destObject, I_32 srcIndex, I_32 destIndex, I_32 lengthInSlots);
+	virtual I_32 forwardReferenceArrayCopyIndex(J9VMThread *vmThread, J9IndexableObject *srcObject, J9IndexableObject *destObject, I_32 srcIndex, I_32 destIndex, I_32 lengthInSlots);
+
+	bool preObjectStoreInternal(J9VMThread *vmThread, J9Object *destClass, J9Object **destAddress, J9Object *value, bool isVolatile);
+	bool preObjectStoreInternal(J9VMThread *vmThread, J9Object *destObject, fj9object_t *destAddress, J9Object *value, bool isVolatile);
+	bool preObjectStoreInternal(J9VMThread *vmThread, J9Object **destAddress, J9Object *value, bool isVolatile);
+
+	MMINLINE bool isBarrierActive(MM_EnvironmentBase* env)
+	{
+		MM_GCExtensions* extensions = MM_GCExtensions::getExtensions(env);
+		return !extensions->sATBBarrierRememberedSet->isGlobalFragmentIndexPreserved(env);
+	}
+
+	MMINLINE bool isDoubleBarrierActiveOnThread(J9VMThread *vmThread)
+	{
+		/* The double barrier is enabled in staccato by setting the threads remembered set fragment index
+		 * to the special value, this ensures the JIT will go out-of line. We can determine if the double
+		 * barrier is active simply by checking if the fragment index corresponds to the special value.
+		 */
+		return (J9GC_REMEMBERED_SET_RESERVED_INDEX == vmThread->sATBBarrierRememberedSetFragment.localFragmentIndex);
+	}
+
+	bool markAndScanContiguousArray(MM_EnvironmentRealtime *env, J9IndexableObject *objectPtr);
+	void scanContiguousArray(MM_EnvironmentRealtime *env, J9IndexableObject *objectPtr);
+
+	/**
+	 * Remember objects that are forced onto the finalizable list at shutdown.
+	 * Called from FinalizerSupport finalizeForcedUnfinalizedToFinalizable
+	 *
+	 * @param vmThread		current J9VMThread (aka JNIEnv)
+	 * @param object		object forced onto the finalizable list
+	 */
+	virtual void forcedToFinalizableObject(J9VMThread* vmThread, J9Object *object);
+
 private:
 	void printClass(J9JavaVM *javaVM, J9Class* clazz);
 };
