--- runtime/gc_staccato/module.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0"?>
-<!--
-Copyright (c) 2007, 2018 IBM Corp. and others
-
-This program and the accompanying materials are made available under
-the terms of the Eclipse Public License 2.0 which accompanies this
-distribution and is available at https://www.eclipse.org/legal/epl-2.0/
-or the Apache License, Version 2.0 which accompanies this distribution and
-is available at https://www.apache.org/licenses/LICENSE-2.0.
-
-This Source Code may also be made available under the following
-Secondary Licenses when the conditions for such availability set
-forth in the Eclipse Public License, v. 2.0 are satisfied: GNU
-General Public License, version 2 with the GNU Classpath
-Exception [1] and GNU General Public License, version 2 with the
-OpenJDK Assembly Exception [2].
-
-[1] https://www.gnu.org/software/classpath/license.html
-[2] http://openjdk.java.net/legal/assembly-exception.html
-
-SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
--->
-<module>
-	<artifact type="static" name="j9staccato">
-		<include-if condition="spec.flags.module_gc_staccato" />
-		<phase>core</phase>
-		<flags>
-			<flag name="-fpeel-loops" asmflag="false" definition="false">
-				<include-if condition="spec.linux_x86.*"/>
-			</flag>
-			<flag name="/w34189" asmflag="false" definition="false">
-				<include-if condition="spec.win_x86.*"/>
-			</flag>
-		</flags>
-		<includes>
-			<include path="j9include"/>
-			<include path="j9oti"/>
-			<include path="j9gcbase"/>
-			<include path="$(OMR_DIR)/gc/base" type="relativepath"/>
-			<include path="$(OMR_DIR)/gc/base/segregated" type="relativepath"/>
-			<include path="$(OMR_DIR)/gc/base/standard" type="relativepath"/>
-			<include path="j9gcstructs"/>
-			<include path="$(OMR_DIR)/gc/structs" type="relativepath"/>
-			<include path="j9gcstats"/>
-			<include path="$(OMR_DIR)/gc/stats" type="relativepath"/>
-			<include path="j9gcinclude"/>
-			<include path="$(OMR_DIR)/gc/include" type="relativepath"/>
-			<include path="j9realtime"/>
-			<include path="j9gcgluejava"/>
-		</includes>
-		<makefilestubs>
-			<makefilestub data="UMA_ENABLE_ALL_WARNINGS=1"/>
-			<makefilestub data="UMA_TREAT_WARNINGS_AS_ERRORS=1"/>
-		</makefilestubs>
-	</artifact>
-</module>
